feat: 439716 - deleteAccount throws 404 if id not found

+++ app/repos/user-accounts.js
@@ -92,6 +92,11 @@ const deleteAccount = async (accountId, user, transaction) => {
     },
     transaction
   })
+
+  if (account === null) {
+    throw new NotFoundError(`Account does not exist with id ${accountId}`)
+  }
+
   await sequelize.models.user_account.destroy({ where: { id: accountId }, transaction })
 
   try {
